Made attributes
Made code attributes to where I got information from. Especially the open CV website which I used to gather the bulk of the video capture code. And also the collaboration with my mentor to get the draw lines code to work and Peteris.rock website for the draw lines solution